Develop

+++ app/Http/Controllers/OfferingTypeController.php
@@ -5,11 +5,12 @@ declare(strict_types=1);
 namespace App\Http\Controllers;
 
 use App\Enums\FlashMessageKey;
-use App\Http\Requests\StoreOfferingTypeRequest;
-use App\Http\Requests\UpdateOfferingTypeRequest;
+use App\Http\Requests\Code\StoreOfferingTypeRequest;
+use App\Http\Requests\Code\UpdateOfferingTypeRequest;
 use App\Http\Resources\Codes\OfferingTypeResource;
 use App\Models\OfferingType;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 use Inertia\Response;
 
@@ -20,6 +21,8 @@ final class OfferingTypeController extends Controller
      */
     public function index(): Response
     {
+        Gate::authorize('viewAny', OfferingType::class);
+
         $offeringTypes = OfferingType::latest()->get();
 
         return Inertia::render('codes/offeringTypes/index', [
@@ -32,9 +35,14 @@ final class OfferingTypeController extends Controller
      */
     public function store(StoreOfferingTypeRequest $request): RedirectResponse
     {
+        Gate::authorize('create', OfferingType::class);
+
         OfferingType::create($request->validated());
 
-        return to_route('codes.offeringTypes.index')->with(FlashMessageKey::SUCCESS->value, 'Offering type created successfully.');
+        return to_route('codes.offeringTypes.index')->with(
+            FlashMessageKey::SUCCESS->value,
+            __('flash.message.created', ['model' => __('Offering Type')])
+        );
     }
 
     /**
@@ -42,9 +50,14 @@ final class OfferingTypeController extends Controller
      */
     public function update(UpdateOfferingTypeRequest $request, OfferingType $offeringType): RedirectResponse
     {
+        Gate::authorize('update', $offeringType);
+
         $offeringType->update($request->validated());
 
-        return to_route('codes.offeringTypes.index')->with(FlashMessageKey::SUCCESS->value, 'Offering type updated successfully.');
+        return to_route('codes.offeringTypes.index')->with(
+            FlashMessageKey::SUCCESS->value,
+            __('flash.message.updated', ['model' => __('Offering Type')])
+        );
     }
 
     /**
@@ -52,8 +65,13 @@ final class OfferingTypeController extends Controller
      */
     public function destroy(OfferingType $offeringType): RedirectResponse
     {
+        Gate::authorize('delete', $offeringType);
+
         $offeringType->delete();
 
-        return to_route('codes.offeringTypes.index')->with(FlashMessageKey::SUCCESS->value, 'Offering type deleted successfully.');
+        return to_route('codes.offeringTypes.index')->with(
+            FlashMessageKey::SUCCESS->value,
+            __('flash.message.deleted', ['model' => __('Offering Type')])
+        );
     }
 }
